Refactored for development in OSX

+++ Strump/Draw.cpp
@@ -359,7 +359,7 @@ void Draw::setSpritePixels() {
 				for (int y = 0; y < 8; y++) {
 					for (int x = 0; x < 8; x++) {
 						getPixel(&cur, x, y, &pixel, sprite->XFlip, sprite->YFlip);
-						uint8_t curPixel = screenPixels[(scY + y) * 160 + (scX + x)];
+						uint32_t curPixel = screenPixels[(scY + y) * 160 + (scX + x)];
 						if (sprite->SpritePriority == 0 || (sprite->SpritePriority == 1 && curPixel != WHITE && curPixel != CLASSIC_WHITE)) {
 							uint32_t colour = GetSpriteColourFor(pixel, sprite->CGBPalette);
 							if(colour != TRANSPARENT)
@@ -378,7 +378,7 @@ void Draw::setSpritePixels() {
 					for (int y = 0; y < 8; y++) {
 						for (int x = 0; x < 8; x++) {
 							getPixel(&cur, x, y, &pixel, sprite->XFlip, sprite->YFlip);
-							uint8_t curPixel = screenPixels[(scY + y) * 160 + (scX + x)];
+							uint32_t curPixel = screenPixels[(scY + y) * 160 + (scX + x)];
 							if (sprite->SpritePriority == 0 || (sprite->SpritePriority == 1 && curPixel != WHITE && curPixel != CLASSIC_WHITE)) {
 								uint32_t colour = GetSpriteColourFor(pixel, sprite->CGBPalette);
 								if (colour != TRANSPARENT)
